Search all, by name and by id in API ordered
