+++ api/src/controllers/booksController.js
@@ -2,16 +2,85 @@ const axios = require ("axios")
 const Book = require('../models/Book')
 
 
+///////////Search in API
+// async function getBooks (req, res){
+//     try {
+//         let booksDefault = []
+//         for (let i = 0; i < 100; i++) {
+//             let data = await axios(`https://api.itbook.store/1.0/search/mongo/${i}`)
+//             booksDefault.push(await data.data.books.map(b => b))
+//         }
+//         let dataDefault = booksDefault.flat()
+//         let apiData = dataDefault.map(b => {
+//             return {
+//                 title: b.title,
+//                 subtitle: b.subtitle,
+//                 isbn13: b.isbn13,
+//                 price: b.price,
+//                 image: b.image
+//             }
+//         })
+//             res.status(200).send(apiData) 
+//     } catch (error) {
+//         res.status(404).json({error: "An unexpected error occurred, please try again later"})
+//     }
+// }
+
+
+// async function getBooksByName (req, res){
+//     const { name } = req.params
+//     try {
+//         let bookSearched = []
+//         for (let i = 0; i < 10; i++) {
+//             let data = await axios(`https://api.itbook.store/1.0/search/${name}/${i}`)
+//             bookSearched.push(await data.data.books.map(b => b))
+//         }
+//         let dataDefault = bookSearched.flat()
+//         let apiData = dataDefault.map(b => {
+//             return {
+//                 title: b.title,
+//                 subtitle: b.subtitle,
+//                 isbn13: b.isbn13,
+//                 price: b.price,
+//                 image: b.image
+//             }
+//         })
+//             res.status(200).send(apiData) 
+//     } catch (error) {
+//         res.status(404).json({error: "An unexpected error occurred, please try again later"})
+//     }
+// }
+
+
+// async function getBooksById (req, res){
+//     const { id } = req.params
+//     try {
+//         let book = await axios(`https://api.itbook.store/1.0/books/${id}`)
+//         delete book.data.error
+//         delete book.data.isbn10
+//         delete book.data.url
+//         res.status(200).send(book.data) 
+//     } catch (error) {
+//         res.status(404).json({error: "An unexpected error occurred, please try again later"})
+//     }
+// }
+
+
+
+///////////Search in DB
+
+
+
 async function getBooks (req, res){
-    // try {
-    //     let dataDefault = await axios(`https://api.itbook.store/1.0/search/mongo`)
-    //         res.status(200).json(dataDefault.data.books)
-    // } catch (error) {
-    //     res.status(404).json({error: "An unexpected error occurred, please try again later"})
-    // }
+    try {
+        let allBooks = await Book.find({})//.select("isbn13")
+        res.status(200).send(allBooks)
+    } catch (error) {
+        res.status(404).json({error: "An unexpected error occurred, please try again later"})
+    }
+}
     // let todo = await Book.find({title: /jira/i})
-    let todo = await Book.find({}).select("isbn13")
-    const tada = todo.map(b => b.isbn13)
+    // const tada = todo.map(b => b.isbn13)
 
     
 //REMOVER DUPLICATEDDDDD
@@ -30,8 +99,6 @@ async function getBooks (req, res){
     // console.log(todo);
 
 
-    res.status(200).send(tada)
-}
 
 
 
@@ -63,38 +130,6 @@ async function getBooks (req, res){
 // }
 
 
-async function getBooksByName (req, res){
-    const { name } = req.params
-    try {
-        let bookSearched = []
-        for (let i = 0; i < 10; i++) {
-            let data = await axios(`https://api.itbook.store/1.0/search/${name}/${i}`)
-            bookSearched.push(await data.data.books.map(b => b))
-            
-        }
-        let dataDefault = bookSearched.flat()
-        let apiData = dataDefault.map(b => {
-            return {
-                title: b.title,
-                subtitle: b.subtitle,
-                isbn13: b.isbn13,
-                price: b.price,
-                image: b.image
-            }
-        })
-        // console.log(apiData)
-        Book.insertMany(apiData, (errors, insertedBooks) => {
-            if(errors) {
-              return res.status(400).json({ok: false, errors});
-            }
-        })
-
-        // const bookSaved = await booksOnDb.save()
-                res.status(201).send("ok") 
-    } catch (error) {
-        res.status(404).json({error: "An unexpected error occurred, please try again later"})
-    }
-}
 
 
 
@@ -104,48 +139,12 @@ async function getBooksByName (req, res){
 
 
 
-async function getBooksById (req, res){
-    // const { id } = req.params
-    try {
-        let todo = await Book.find({}).select("isbn13")
-        const tada = todo.map(b => b.isbn13)
-        let array = []
-        const arrayNoexiste = []
-        // array.push(tada[0])
-        const elMap = tada.map(async (doc) => {
-            // console.log("DOCCCCCCC", typeof doc);
-            // setTimeout(() => {}, 2000)
-            // var url = "https://api.itbook.store/1.0/books/" + doc
-            // console.log(url);
-            var book = await axios.get("https://api.itbook.store/1.0/books/" + doc)
-            .catch((err) => {
-                arrayNoexiste.push(doc)
-                console.log(arrayNoexiste)
-                return null
-            } )
-            if (!book){
-                // arrayNoexiste.push(doc)
-                return null
-            }
-            array.push(arrayNoexiste)
-            // console.log(book);
-        // console.log(book.data);
-        })
-        
-        
-        // let book = await axios(`https://api.itbook.store/1.0/books/${id}`)
-        res.status(200).send(elMap) 
-    } catch (error) {
-        res.status(404).json({error: "An unexpected error occurred, please try again later"})
-    }
-    // console.log(arrayNoexiste, "b")
-}
 
 
 
 
 module.exports = {
     getBooks,
-    getBooksByName,
-    getBooksById
+    // getBooksByName,
+    // getBooksById
 }
