+++ api/src/routes/books.js
@@ -1,5 +1,5 @@
 const { Router } = require ('express')
-const { getBooks, getBooksByName, getBooksById } = require('../controllers/booksController')
+const { getBooks } = require('../controllers/booksController')
 
 const router = Router()
 
@@ -7,6 +7,6 @@ router.get("/", getBooks)
 
 // router.get("/:name", getBooksByName)
 
-router.get("/id", getBooksById)
+// router.get("/id/:id", getBooksById)
 
 module.exports = router
